@@ -21,29 +22,29 @@
     "@uppy/xhr-upload": "^4.3.3",
     "antd": "^5.13.3",
     "axios": "^1.6.7",
-    "babel-jest": "^29.7.0",
+    "babel-jest": "^30.0.0-beta.3",
     "classnames": "^2.5.1",
-    "css-loader": "^6.10.0",
+    "css-loader": "^7.1.2",
     "dayjs": "^1.11.10",
     "debounce": "^2.2.0",
     "dotenv": "^16.4.1",
-    "eslint": "^8.56.0",
+    "eslint": "^9.28.0",
     "file-loader": "^6.2.0",
     "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
+    "jest-environment-jsdom": "^30.0.0-beta.3",
     "jest-transform-stub": "^2.0.0",
     "lodash.throttle": "^4.1.1",
     "moment": "^2.30.1",
-    "ramda": "^0.29.1",
-    "react": "^18.2.0",
+    "ramda": "^0.30.1",
+    "react": "^19.1.0",
     "react-dev-utils": "^12.0.1",
-    "react-dom": "^18.2.0",
+    "react-dom": "^19.1.0",
     "react-image-gallery": "^1.3.0",
     "react-redux": "^9.1.0",
-    "react-router-dom": "^5.00.0",
+    "react-router-dom": "^7.6.2",
     "redux-thunk": "^3.1.0",
     "reselect": "^5.1.0",
-    "uuid": "^9.0.1",
+    "uuid": "^11.1.0",
     "video.js": "^8.21.0",
     "workbox-webpack-plugin": "^7.0.0"
   },
@@ -83,47 +84,43 @@
     "@babel/preset-env": "^7.23.9",
     "@babel/preset-react": "^7.23.3",
     "@babel/preset-typescript": "^7.23.3",
-    "@rsbuild/core": "^0.7.10",
-    "@rspack/cli": "^0.7.5",
-    "@rspack/core": "^0.7.5",
-    "@rspack/plugin-react-refresh": "^0.7.5",
-    "@testing-library/dom": "^9.3.4",
+    "@rsbuild/core": "^1.3.22",
+    "@rspack/cli": "^1.3.14",
+    "@rspack/core": "^1.3.14",
+    "@rspack/plugin-react-refresh": "^1.4.3",
+    "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.4.1",
-    "@testing-library/react": "^14.2.0",
+    "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.11",
     "@types/lodash.throttle": "^4.1.9",
-    "@types/node": "^20.11.15",
-    "@types/ramda": "^0.29.10",
-    "@types/react": "^18.2.51",
-    "@types/react-dom": "^18.2.18",
+    "@types/node": "^22.15.30",
+    "@types/ramda": "^0.30.2",
+    "@types/react": "^19.1.6",
+    "@types/react-dom": "^19.1.6",
     "@types/react-image-gallery": "^1.2.4",
     "@types/react-router-dom": "^5.3.3",
-    "@types/uuid": "^9.0.8",
+    "@types/uuid": "^10.0.0",
     "@types/video.js": "^7.3.58",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
-    "@typescript-eslint/parser": "^6.20.0",
+    "@typescript-eslint/eslint-plugin": "^8.33.1",
+    "@typescript-eslint/parser": "^8.33.1",
     "autoprefixer": "^10.4.17",
     "cross-env": "^7.0.3",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-standard-react": "^13.0.0",
-    "eslint-import-resolver-webpack": "^0.13.8",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-unicorn": "^50.0.1",
-    "exiftool-vendored": "^26.0.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "exiftool-vendored": "^30.2.0",
     "gh-pages": "^6.1.1",
-    "husky": "^8.0.0",
-    "lint-staged": "^15.2.1",
-    "react-refresh": "^0.14.2",
+    "husky": "^9.1.7",
+    "lint-staged": "^16.1.0",
+    "react-refresh": "^0.17.0",
     "sass": "^1.70.0",
     "sass-embedded": "^1.77.5",
-    "sass-loader": "^14.2.1",
+    "sass-loader": "^16.0.5",
     "ts-jest": "^29.1.2",
-    "typescript": "5.3.3",
+    "typescript": "5.8.3",
+    "typescript-eslint": "^8.33.1",
     "typescript-plugin-css-modules": "^5.0.2",
     "webpack-bundle-analyzer": "^4.10.1",
     "webpack-merge": "^6.0.1"
